+++ packages/assets-controllers/src/TokenListController.ts
@@ -68,12 +68,14 @@ export type TokenListControllerActions = GetTokenListState;
 
 type AllowedActions = NetworkControllerGetNetworkClientByIdAction;
 
-type TokenListMessenger = RestrictedControllerMessenger<
+type AllowedEvents = NetworkControllerStateChangeEvent;
+
+export type TokenListControllerMessenger = RestrictedControllerMessenger<
   typeof name,
   TokenListControllerActions | AllowedActions,
-  TokenListControllerEvents | NetworkControllerStateChangeEvent,
+  TokenListControllerEvents | AllowedEvents,
   AllowedActions['type'],
-  (TokenListControllerEvents | NetworkControllerStateChangeEvent)['type']
+  AllowedEvents['type']
 >;
 
 const metadata = {
@@ -82,10 +84,12 @@ const metadata = {
   preventPollingOnNetworkRestart: { persist: true, anonymous: true },
 };
 
-const defaultState: TokenListState = {
-  tokenList: {},
-  tokensChainsCache: {},
-  preventPollingOnNetworkRestart: false,
+export const getDefaultTokenListState = (): TokenListState => {
+  return {
+    tokenList: {},
+    tokensChainsCache: {},
+    preventPollingOnNetworkRestart: false,
+  };
 };
 
 /**
@@ -94,7 +98,7 @@ const defaultState: TokenListState = {
 export class TokenListController extends StaticIntervalPollingController<
   typeof name,
   TokenListState,
-  TokenListMessenger
+  TokenListControllerMessenger
 > {
   private readonly mutex = new Mutex();
 
@@ -136,14 +140,14 @@ export class TokenListController extends StaticIntervalPollingController<
     ) => void;
     interval?: number;
     cacheRefreshThreshold?: number;
-    messenger: TokenListMessenger;
+    messenger: TokenListControllerMessenger;
     state?: Partial<TokenListState>;
   }) {
     super({
       name,
       metadata,
       messenger,
-      state: { ...defaultState, ...state },
+      state: { ...getDefaultTokenListState(), ...state },
     });
     this.intervalDelay = interval;
     this.cacheRefreshThreshold = cacheRefreshThreshold;