+++ packages/assets-controllers/src/index.ts
@@ -4,7 +4,14 @@ export * from './CurrencyRateController';
 export * from './NftController';
 export * from './NftDetectionController';
 export * from './TokenBalancesController';
-export * from './TokenDetectionController';
+export type {
+  TokenDetectionControllerMessenger,
+  TokenDetectionControllerActions,
+  TokenDetectionControllerGetStateAction,
+  TokenDetectionControllerEvents,
+  TokenDetectionControllerStateChangeEvent,
+} from './TokenDetectionController';
+export { TokenDetectionController } from './TokenDetectionController';
 export * from './TokenListController';
 export * from './TokenRatesController';
 export * from './TokensController';
@@ -12,5 +19,6 @@ export {
   isTokenDetectionSupportedForNetwork,
   formatIconUrlWithProxy,
   getFormattedIpfsUrl,
+  fetchTokenContractExchangeRates,
 } from './assetsUtil';
 export { CodefiTokenPricesServiceV2 } from './token-prices-service';